Reduce oscap info output
After we added SCE checks to ComplianceAsCode data streams we discovered
that the oscap info output is appalling because it's polluted by all the
SCE check information. In this commit, we will address this problem by
removing SCE checks from the "Referenced check files:" section. Then, we
will completely remove the "Checks:" and "Dictionaries:" sections
because they only list crefs and crefs are internal data stream
identifiers that noone should be interested in.