Rename oscap-bootc to oscap-im
This commit renames the `oscap-bootc` tool to `oscap-im`. IM stands for
Image Mode. The `bootc` is just one of the tools that participate on the
Image Mode Operating System experience, therefore, it's more appropriate
to use Image Mode than bootc in the tooling naming. It should be safe to
rename it now because the `oscap-bootc` hasn't been released in upstream
nor in downstream.