OPENSCAP-5235: Block remediation on deployed bootc system
OpenSCAP remediation is supposed to be used only at bootc container
image build. Deployed bootc system is immutable, it can't be remediated
with OpenSCAP and trying to do so would result in errors and bad user
experience.

We will update OpenSCAP to print error message for users in case they
try to run remediation on an already deployed bootc system, informing
them that it is not possible and that the openscap remediation must be
performed during container build.